benchmarks/requirements-bench.txt
numpy>=1.24
scipy>=1.10
pandas>=2.0
psutil>=5.9
networkx>=3.0
scikit-learn>=1.3
tqdm>=4.66
torch>=2.1
pennylane>=0.36
qandle
pytest-benchmark>=4.0
